+++ StorageDLHI.App/StorageDLHI.DAL/Models/Pos.cs
@@ -23,7 +23,7 @@ namespace StorageDLHI.DAL.Models
         public string Po_Approved { get; set; }
         public string Po_Payment_Term { get; set; }
         public string Po_Dispatch_Box { get; set; }
-        public Int64 Po_Total_Amount { get; set; }
+        public double Po_Total_Amount { get; set; }
         public bool IsImported { get; set; }
         
         public Guid CostId { get; set; }
